requirements.txt
telethon==1.36.0
ccxt==4.3.73
openai==1.51.0
fastapi==0.115.0
uvicorn[standard]==0.31.0
cryptography==43.0.1
python-dotenv==1.0.1
pydantic==2.9.2
httpx==0.27.2
